requirements.txt
gspread
google-auth
Pillow
requests
